vc demo upgrade and renaming
