+++ script/tsconfig.json
@@ -14,7 +14,7 @@
     "declaration": true
   },
   "include": [
-    "src/**/*"
+    "./**/*"
   ],
   "exclude": [
     "node_modules"
