+++ src/identity/hfs/vc-status-list/vc-sl-status.ts
@@ -1,4 +1,4 @@
-export enum VcSlStatuses {
+export enum VcSlStatus {
     ACTIVE = 0,
     RESUMED = 1,
     SUSPENDED = 2,
