+++ test/revocation-list.spec.js
@@ -1,23 +1,24 @@
+/*!
+ * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
+ */
 const { PrivateKey, Client, FileContentsQuery } = require("@hashgraph/sdk");
-const { HfsVcSl, VcSlStatus } = require("../../dist");
-const { OPERATOR_ID, OPERATOR_KEY } = require("../.env.json");
+const { HfsVcSl, VcSlStatus } = require("../dist");
+const { OPERATOR_ID, OPERATOR_KEY } = require("./.env.json");
 
 describe("HfsVcSl", () => {
-    let hfsVcSl;
+    let hcsRl;
     let client;
     let fileId;
 
     before(async () => {
         client = Client.forTestnet();
         client.setOperator(OPERATOR_ID, OPERATOR_KEY);
-        const revocationListOwnerPrivateKey = PrivateKey.generate();
 
-        hfsVcSl = new HfsVcSl(
+        hcsRl = new HfsVcSl(
             PrivateKey.fromString(OPERATOR_KEY), // this is to sign transaction
-            client,
-            revocationListOwnerPrivateKey
+            client
         );
-        fileId = await hfsVcSl.createRevocationListFile();
+        fileId = await hcsRl.createRevocationListFile();
     });
 
     describe("revocation list", () => {
@@ -34,27 +35,27 @@ describe("HfsVcSl", () => {
 
         describe("#CredentialStatus", () => {
             it("should apply revoke status to revocation list index 0", async () => {
-                const test = await hfsVcSl.revokeByIndex(fileId, 0);
-                const status = await hfsVcSl.resolveStatusByIndex(fileId, 0);
-                assert.equal(VcSlStatus[status], VcSlStatus.REVOKED);
+                await hcsRl.revokeByIndex(fileId, 0);
+                const status = await hcsRl.resolveStatusByIndex(fileId, 0);
+                assert.equal(VcSlStatus[status], VcSlStatus.REVOKE);
             });
 
             it("should apply suspend status to revocation list index 0", async () => {
-                await hfsVcSl.suspendByIndex(fileId, 0);
-                const status = await hfsVcSl.resolveStatusByIndex(fileId, 0);
+                await hcsRl.suspendByIndex(fileId, 0);
+                const status = await hcsRl.resolveStatusByIndex(fileId, 0);
                 assert.equal(VcSlStatus[status], VcSlStatus.SUSPENDED);
             });
 
             it("should apply resume status to revocation list index 0", async () => {
-                await hfsVcSl.resumeByIndex(fileId, 0);
-                const status = await hfsVcSl.resolveStatusByIndex(fileId, 0);
-                assert.equal(VcSlStatus[status], VcSlStatus.RESUMED);
+                await hcsRl.resumeByIndex(fileId, 0);
+                const status = await hcsRl.resolveStatusByIndex(fileId, 0);
+                assert.equal(VcSlStatus[status], VcSlStatus.RESUME);
             });
 
             it("should apply issue status to revocation list index 0", async () => {
-                await hfsVcSl.issueByIndex(fileId, 0);
-                const status = await hfsVcSl.resolveStatusByIndex(fileId, 0);
-                assert.equal(VcSlStatus[status], VcSlStatus.ACTIVE);
+                await hcsRl.issueByIndex(fileId, 0);
+                const status = await hcsRl.resolveStatusByIndex(fileId, 0);
+                assert.equal(VcSlStatus[status], VcSlStatus.ISSUE);
             });
         });
     });
